Feature/17 json graph io managment (#52)
* Json reader v1

* Updated the json reader

* Write json graph reader

* Json graph reader fixed for logger

* Moved io files to folders

* Json graph reader update V2

* Written tests

* Changed graph vertex colors to be uint32 instead of int32

* Fixed includes

* Fixed documintation

* Fixed type comperesent

* Fixed exception type thrown in file logger

* Fixed catching exception in more specific spots

* Formatted everything

* Updated cmake because files moved from io to io/graph

* Add boost::json to the cmake requierments and linkage

* Fixed exception handeling

* Formatted code

* Started updating tests - not working yet

+++ src/graph/ColoredGraph.cpp
@@ -220,7 +220,7 @@ ColoredGraph::compute_range(uint32_t vertex, const std::vector<uint32_t>& elemen
     const std::vector<uint32_t>::const_iterator range_begin =
         elements.begin() + index_of_neighbours.at(vertex);
     const std::vector<uint32_t>::const_iterator range_end =
-        (vertex + 1 < index_of_neighbours.size())
+        (static_cast<size_t>(vertex) + 1UL < index_of_neighbours.size())
             ? elements.begin() + index_of_neighbours.at(vertex + 1)
             : elements.end();
     return {range_begin, range_end};
